Move items details to detail folder
