+++ src/components/Detail/ItemsDetail/component.tsx
