+++ src/components/Detail/ItemsDetail/index.tsx
