+++ src/components/Detail/ItemsDetail/interface.tsx
