+++ src/components/Detail/ItemsDetail/styles.module.scss
